Move form into seperate page
