+++ frontend/src/PagesAdmin/RecruitmentPositionFormAdminPage/RecruitmentPositionFormAdminPage.tsx
@@ -1,33 +1,15 @@
-import { zodResolver } from '@hookform/resolvers/zod';
 import { useEffect, useState } from 'react';
-import { Controller, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
-import type { z } from 'zod';
-import {
-  Button,
-  Checkbox,
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-  Input,
-  Textarea,
-} from '~/Components';
-import { getRecruitmentPosition, postRecruitmentPosition, putRecruitmentPosition } from '~/api';
+import { getRecruitmentPosition } from '~/api';
 import type { RecruitmentPositionDto } from '~/dto';
 import { useTitle } from '~/hooks';
 import { KEY } from '~/i18n/constants';
 import { reverse } from '~/named-urls';
 import { ROUTES } from '~/routes';
 import { AdminPageLayout } from '../AdminPageLayout/AdminPageLayout';
-import styles from './RecruitmentPositionFormAdminPage.module.scss';
-import { recruitmentPositionSchema } from './RecruitmentPositionFormSchema';
-
-type RecruitmentPositionFormType = z.infer<typeof recruitmentPositionSchema>;
+import { RecruitmentPositionForm } from './RecruitmentPositionForm';
 
 export function RecruitmentPositionFormAdminPage() {
   const { t } = useTranslation();
@@ -54,7 +36,7 @@ export function RecruitmentPositionFormAdminPage() {
     }
   }, [positionId, recruitmentId, gangId, navigate, t]);
 
-  const initialData: Partial<RecruitmentPositionFormType> = {
+  const initialData: Partial<RecruitmentPositionDto> = {
     name_nb: position?.name_nb || '',
     name_en: position?.name_en || '',
     norwegian_applicants_only: position?.norwegian_applicants_only || false,
@@ -68,219 +50,20 @@ export function RecruitmentPositionFormAdminPage() {
     tags: position?.tags || '',
   };
 
-  const form = useForm<RecruitmentPositionFormType>({
-    resolver: zodResolver(recruitmentPositionSchema),
-    defaultValues: initialData,
-  });
-
-  useEffect(() => {
-    if (position) {
-      form.reset(position);
-    }
-  }, [position, form]);
-
   const title = positionId
     ? `${t(KEY.common_edit)} ${position?.name_nb}`
     : `${t(KEY.common_create)} ${t(KEY.recruitment_position)}`;
 
   useTitle(title);
 
-  const submitText = positionId ? t(KEY.common_save) : t(KEY.common_create);
-
-  function onSubmit(data: RecruitmentPositionFormType) {
-    const updatedPosition = {
-      ...data,
-      gang: { id: Number.parseInt(gangId ?? '') },
-      recruitment: recruitmentId ?? '',
-      interviewers: [],
-    };
-
-    const action = positionId
-      ? putRecruitmentPosition(positionId, updatedPosition)
-      : postRecruitmentPosition(updatedPosition);
-
-    action
-      .then(() => {
-        toast.success(positionId ? t(KEY.common_update_successful) : t(KEY.common_creation_successful));
-        navigate(
-          reverse({
-            pattern: ROUTES.frontend.admin_recruitment_gang_position_overview,
-            urlParams: { recruitmentId, gangId },
-          }),
-        );
-      })
-      .catch((error) => {
-        toast.error(t(KEY.common_something_went_wrong));
-        console.error(error);
-      });
-  }
-
   return (
     <AdminPageLayout title={title} header={true}>
-      <Form {...form}>
-        <form onSubmit={form.handleSubmit(onSubmit)} className={styles.form}>
-          <div className={styles.wrapper}>
-            <Controller
-              name="norwegian_applicants_only"
-              control={form.control}
-              render={({ field }) => (
-                <FormItem className={styles.item}>
-                  <FormLabel>{t(KEY.recruitment_norwegian_applicants_only)}</FormLabel>
-                  <FormControl>
-                    <Checkbox checked={field.value} onChange={field.onChange} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <div className={styles.row}>
-              <FormField
-                control={form.control}
-                name="name_nb"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.common_name)} ${t(KEY.common_norwegian)}`}</FormLabel>
-                    <FormControl>
-                      <Input {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name="name_en"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.common_name)} ${t(KEY.common_english)}`}</FormLabel>
-                    <FormControl>
-                      <Input {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-
-            <div className={styles.row}>
-              <FormField
-                control={form.control}
-                name="short_description_nb"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.common_short_description)} ${t(KEY.common_norwegian)}`}</FormLabel>
-                    <FormControl>
-                      <Input type="text" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name="short_description_en"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.common_short_description)} ${t(KEY.common_english)}`}</FormLabel>
-                    <FormControl>
-                      <Input type="text" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-
-            <div className={styles.row}>
-              <FormField
-                control={form.control}
-                name="long_description_nb"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.common_long_description)} ${t(KEY.common_norwegian)}`}</FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name="long_description_en"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.common_long_description)} ${t(KEY.common_english)}`}</FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-
-            <Controller
-              name="is_funksjonaer_position"
-              control={form.control}
-              render={({ field }) => (
-                <FormItem className={styles.item}>
-                  <FormLabel>{t(KEY.recruitment_funksjonaer)}</FormLabel>
-                  <FormControl>
-                    <Checkbox checked={field.value} onChange={field.onChange} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-
-            <div className={styles.row}>
-              <FormField
-                control={form.control}
-                name="default_application_letter_nb"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.recrutment_default_application_letter)} ${t(KEY.common_norwegian)}`}</FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name="default_application_letter_en"
-                render={({ field }) => (
-                  <FormItem className={styles.item}>
-                    <FormLabel>{`${t(KEY.recrutment_default_application_letter)} ${t(KEY.common_english)}`}</FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-
-            <FormField
-              control={form.control}
-              name="tags"
-              render={({ field }) => (
-                <FormItem className={styles.item}>
-                  <FormLabel>{t(KEY.common_tags)}</FormLabel>
-                  <FormControl>
-                    <Input {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button type="submit">{submitText}</Button>
-          </div>
-        </form>
-      </Form>
+      <RecruitmentPositionForm
+        initialData={initialData}
+        positionId={positionId}
+        recruitmentId={recruitmentId}
+        gangId={gangId}
+      />
     </AdminPageLayout>
   );
 }
